Dependencies
Had to collect a few libraries to program off of. I have Wad.js which
will do the DAW, I have Recorder.js which will record and then download
the file, I have Bootstrap for CSS templating, I have jQuery because
glue is a must, and I have a basic script for file uploading that I
pulled from the internet. All of these will create the app as needed.